ADD support for saml2 REMOVE ldap support
Convert from LDAP to using the SimpleSAMLphp extension for auth. This is
fall out from the attempt to block basic CWL users. Turns out lack of a
PUID doesn't indicate basic CWL, but rather that we didn't have LDAP
permission for a whole slew of attributes. Long story short, due to the
fact that LDAP doesn't have MFA, continuing to use it will probably
make the security folks angry. So we're now trying to switch to
SAML2/Shibboleth SSO.

Blocking basic CWL on our side might not be necessary anymore, as
apparently, the IDP can exclude such accounts for us. But given that we
would still like to retain the ability for non-UBC collaborators to edit
the wiki, I'm not sure we want to exclude at the IDP. So I've retained
the blocking code. This gives us the option of admins manually
unblocking legit basic CWL users.

The 'eduPersonAffiliation' attribute should tell us what kind of user
we're dealing with (staff, student, faculty, etc). Unfortunately, we're
not sure what value would be in there for a basic CWL user. Currently,
we're assuming that it'll be empty. This'll probably change.

The UsernameInfoProvider is the equivalent of LDAP extension's username
normalizer. It is plugged in via config in LocalSettings.php like so:

  $wgSimpleSAMLphp_MandatoryUserInfoProviders['username'] = [
      'class' => 'MediaWiki\Extension\UBCAuth\UsernameInfoProvider'
  ];

It takes over the duties of the ldapUserToCWL() function.

It seems that every SAML2 attribute is treated as an array of values. So
I added a helper getStringAttr() that'll assume this is a supposed to be
a single value string and return just the first value in the array.
getArrayAttr() can be used to return the array as is.

There is a UserNameUtils in the API that might be helpful as a check to
reject invalid usernames, but it unfortunately doesn't do sanitization.
So I'm leaving the old code in there to save us time for now, but might
be good investigate UserNameUtils for later.

Looks like we're in a transition period for the Mediawiki DB API. I've
updated some calls and added "TODO" for some of them that'll need to be
updated when we upgrade >= 1.42 or so.