ADD update cwl extended account data on login
Added a onUserLoggedIn hook that'll check if the user cwl extended
account data (ucead) needs to be updated on every login. This is in the
hopes that we'll see recent legit accounts get their ubc affiliation
filled out when they visit the wiki. This will give us more confidence
that they're not bots. Older accounts have their affiliation filled out,
but at some point, the LDAP account lost permission to those attribute,
so we've been getting blank data for newer accounts.

The update really only applies to the puid, full name (CWLNickname
column in the table), and ubc affiliation (CWLRole column). If the
SAML2 attributes don't match what we have stored, it'll update the
stored values using the SAML2 attributes.

The getUceadByCwlLogin() static helper in Hooks is meant to grab the row
in the ucead table corresponding to the provided CWL login name. This
helper is used by the new _update_cwl_extended_account_data() function
and the UsernameProvider.

The UsernameProvider used to only pass the cwl data along if it's a new
user. Wih the addition of the update function, it now needs to pass the
cwl data along for both new and existing users.

The _get_cwl_data() helper is a refactor, as both create and update cwl
extended data functions needs to do the same thing. I also moved the
data sanitization/conversion into it.