FIX new users cannot login due to update cwl data
For new users, they already read out and cleared the cwl data with the
_create_cwl_extended_account_data() call. So when
_update_cwl_extended_account_data() happens later with the
onUserLoggedIn() hook, we get can exception that the CWL data is empty.

To fix this, I moved the empty cwl account exception to only happen with
_create_cwl_extended_account_data() since that's the critical one. For
the _update_cwl_extended_account_data(), if there's no cwl data saved in
the auth manager, it'll just skip updating.