EDIT autoblocker login check, skip unblocked users
Added autoblocker check on every login. This is to catch all potentially
spam basic CWL accounts that existed before we added the autoblocker.
Such accounts will require an admin to manually unblock them.

The autoblocker on login is called by the
_update_cwl_extended_account_data(). Not the best place to put it, just
convenient because it has access to all the data we need. It checks the
new past role field instead of current/new roles.

The new past role field (CWLRolePast) is meant to avoid sponsored CWL
users from getting blocked when their sponsorship (max 13 months)
expires. The CWLRolePast is meant to contain every role we've seen for
the given user.

For accounts where an admin unblocked them, we add an 'unblocked' past
role to them so that they won't get caught in the autoblocker again.

CWLNickname sanitization was removed as the Shibboleth attribute it
comes from should already be sanitized.

New database migration was added to add the CWLRolePast field to the
user_cwl_extended_account_data table. I dropped the unused fields such
as ubc_role_id and ubc_dept_id. Kept account_status even though it's
currently unused as it might be potentially useful. I found out that
varchar is being converted into varbinary by mediawiki, as they don't
want to deal with collation issues. So I also added a migration to
varbinary for the old fields, also increased the length of the old
fields since that seems useful. Also added an index on CWLLogin since
every login will search for the ucead entry based on CWLLogin.

On varchar, indexing is faster if you limit to 255, I'm not sure if that
applies to varbinary, but I've limited CWLLogin to 255 too just in case.
I can see searching by CWLRole, but for CWLNickname and CWLRolePast, I
think it's fine to let them be larger, hence the larger limit.

Mediawiki maintainence/update.php must be run for this update.