EDIT block reason now uses wiki message system
The wiki localisation system is now used to handle the block reason
message. The i18n/qqq.json stores documentation for each message key.
While i18n/en.json stores the actual message. The message is now
customizable by admins by editing
MediaWiki:Ubcauth-basiccwl-block-reason

The hardcoded reason would prove difficult to modify as we'd have to
deploy an entirely new wiki image just to change the message. Hence the
integration with the localisation system. This came about because of a
request to add an email address to the message.